+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+1.1.0 (2024-05-12)
+------------------
+* Added field `Cv`: Isochoric heat capacity
+* Added field `comp`: Isothermal compressibility
+
 1.0.0 (2024-05-11)
 ------------------
 * Initial release